HSEARCH-3273 Keep LuceneIndexFieldType instances at runtime
Instead of just copying their content to the schema node.

We'll need to keep them at runtime anyway if we want to implement
templates.
