+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/document/impl/AbstractLuceneDocumentBuilder.java
@@ -51,7 +51,7 @@ abstract class AbstractLuceneDocumentBuilder implements LuceneDocumentBuilder {
 			checkNoValueYetForSingleValued( fieldSchemaNode.getAbsoluteFieldPath() );
 		}
 
-		fieldSchemaNode.getCodec().encode( this, fieldSchemaNode.getAbsoluteFieldPath(), value );
+		fieldSchemaNode.getType().getCodec().encode( this, fieldSchemaNode.getAbsoluteFieldPath(), value );
 	}
 
 	@Override
