+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/search/aggregation/impl/LuceneSearchAggregationBuilderFactory.java
@@ -105,7 +105,7 @@ public class LuceneSearchAggregationBuilderFactory
 
 		@Override
 		public LuceneFieldAggregationBuilderFactory extractComponent(LuceneIndexSchemaFieldNode<?> schemaNode) {
-			return schemaNode.getAggregationBuilderFactory();
+			return schemaNode.getType().getAggregationBuilderFactory();
 		}
 
 		@Override
